+++ NEWS.md
@@ -1,8 +1,19 @@
 # tidyr 0.8.2.9000
 
+* `unnest()` preserves column being unnested when input is zero-length (#483).
+
+* `unnest()` will run with named and unnamed list-columns of same length
+  (@hlendway, #460).
+
+* `crossing()` (and hence `expand()`) now works with list-columns 
+  (#446, @SamanthaToet).
+
 * `crossing()` preserves factor levels (#410).
 
 * `gather` help page includes @apreshill's GIF. (@khailper #515)
+
+* `spread()` works when the id variable has names (#525).
+
 
 # tidyr 0.8.2